requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
exceptiongroup==1.2.2
Flask==2.3.2
Flask-Cors==5.0.0
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
ibm-cloud-sdk-core==3.21.0
ibm-watson==8.1.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.3
langchain-core==0.3.10
langchain-text-splitters==0.3.0
langsmith==0.1.134
MarkupSafe==3.0.1
multidict==6.1.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pluggy==1.5.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==7.3.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==8.5.0
tomli==2.0.2
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
yarl==1.14.0
